remove internal pkg and move base code to certain place
