+++ pkg/planner/core/BUILD.bazel
@@ -113,9 +113,8 @@ go_library(
         "//pkg/planner/cardinality",
         "//pkg/planner/context",
         "//pkg/planner/core/base",
-        "//pkg/planner/core/internal",
-        "//pkg/planner/core/internal/base",
         "//pkg/planner/core/metrics",
+        "//pkg/planner/core/operator/base",
         "//pkg/planner/funcdep",
         "//pkg/planner/property",
         "//pkg/planner/util",
@@ -251,9 +250,9 @@ go_test(
         "//pkg/parser/terror",
         "//pkg/planner",
         "//pkg/planner/core/base",
-        "//pkg/planner/core/internal",
         "//pkg/planner/property",
         "//pkg/planner/util",
+        "//pkg/planner/util/coretestsdk",
         "//pkg/planner/util/coreusage",
         "//pkg/session",
         "//pkg/session/types",
