+++ pkg/planner/core/casetest/enforcempp/BUILD.bazel
@@ -13,7 +13,7 @@ go_test(
     deps = [
         "//pkg/domain",
         "//pkg/parser/model",
-        "//pkg/planner/core/internal",
+        "//pkg/planner/util/coretestsdk",
         "//pkg/sessionctx/stmtctx",
         "//pkg/testkit",
         "//pkg/testkit/external",
