+++ pkg/planner/core/casetest/windows/BUILD.bazel
@@ -13,7 +13,7 @@ go_test(
     shard_count = 5,
     deps = [
         "//pkg/domain",
-        "//pkg/planner/core/internal",
+        "//pkg/planner/util/coretestsdk",
         "//pkg/testkit",
         "//pkg/testkit/testdata",
         "//pkg/testkit/testmain",
