+++ pkg/planner/core/operator/base/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "base",
     srcs = ["plan.go"],
-    importpath = "github.com/pingcap/tidb/pkg/planner/core/internal/base",
+    importpath = "github.com/pingcap/tidb/pkg/planner/core/operator/base",
     visibility = ["//pkg/planner/core:__subpackages__"],
     deps = [
         "//pkg/expression",
