+++ pkg/planner/util/coretestsdk/BUILD.bazel
@@ -1,17 +1,12 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "internal",
-    srcs = [
-        "testkit.go",
-        "util.go",
-    ],
-    importpath = "github.com/pingcap/tidb/pkg/planner/core/internal",
-    visibility = ["//pkg/planner/core:__subpackages__"],
+    name = "coretestsdk",
+    srcs = ["testkit.go"],
+    importpath = "github.com/pingcap/tidb/pkg/planner/util/coretestsdk",
+    visibility = ["//visibility:public"],
     deps = [
         "//pkg/domain",
-        "//pkg/expression",
-        "//pkg/expression/aggregation",
         "//pkg/parser/model",
         "//pkg/store/mockstore",
         "//pkg/store/mockstore/unistore",
