+++ pkg/planner/util/coreusage/BUILD.bazel
@@ -3,10 +3,15 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "coreusage",
     srcs = [
+        "cast_misc.go",
         "cost_misc.go",
         "opt_tracer.go",
     ],
     importpath = "github.com/pingcap/tidb/pkg/planner/util/coreusage",
     visibility = ["//visibility:public"],
-    deps = ["//pkg/util/tracing"],
+    deps = [
+        "//pkg/expression",
+        "//pkg/expression/aggregation",
+        "//pkg/util/tracing",
+    ],
 )
