+++ pkg/planner/util/coreusage/cast_misc.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package internal
+package coreusage
 
 import (
 	"github.com/pingcap/tidb/pkg/expression"
